# src/PaperDotNet.Host/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "paperdotnet | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5098"
    }
  ],
  "paths": {
...
    "/v1.0/me": {
      "get": {
        "tags": [
          "Me"
        ],
        "operationId": "GetMe",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Me"
        ],
        "operationId": "UpdateMe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1.0/me/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "description": "Newest first; ?unreadOnly=true filters.",
        "operationId": "ListNotifications",
        "parameters": [
          {
            "name": "$top",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "$skiptoken",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unreadOnly",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfNotificationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1.0/me/notifications/{id}/read": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "operationId": "MarkNotificationRead",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/v1.0/me/notifications/{id}": {
      "delete": {
        "tags": [
          "Notifications"
        ],
        "operationId": "DeleteNotification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/v1.0/me/notificationSettings": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "operationId": "GetNotificationSettings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Notifications"
        ],
        "description": "A new webhook URL creates a signing secret, returned once in webhookSecret. Quiet hours and the digest hour are local times in the user's time zone.",
        "operationId": "ReplaceNotificationSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1.0/me/subscriptions": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "operationId": "ListSubscriptions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FollowResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Notifications"
        ],
        "description": "Follows a list or an item the caller can read; following again changes the frequency.",
        "operationId": "CreateSubscription",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1.0/me/subscriptions/{id}": {
      "delete": {
        "tags": [
          "Notifications"
        ],
        "operationId": "DeleteSubscription",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChannelChoice": {
        "type": "object",
        "properties": {
          "inApp": {
            "type": "boolean",
            "default": true
          },
          "webhook": {
            "type": "boolean",
            "default": true
          }
        },
        "description": "Channel choice for one notification type."
      },
...
      "FollowRequest": {
        "required": [
          "workspaceId",
          "listId",
          "itemId",
          "frequency"
        ],
        "type": "object",
        "properties": {
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "listId": {
            "type": "string",
            "format": "uuid"
          },
          "itemId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "frequency": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "Follow body; `frequency` is `immediate` (default) or `daily`."
      },
      "FollowResponse": {
        "required": [
          "id",
          "workspaceId",
          "listId",
          "itemId",
          "frequency",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "listId": {
            "type": "string",
            "format": "uuid"
          },
          "itemId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "frequency": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "MeResponse": {
        "required": [
          "id",
          "userName",
          "displayName",
          "email",
          "tenantId",
          "tenantIdentifier",
          "scopes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "tenantIdentifier": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "NotificationResponse": {
        "required": [
          "id",
          "type",
          "title",
          "body",
          "workspaceId",
          "listId",
          "itemId",
          "createdAt",
          "readAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": [
              "null",
              "string"
            ]
          },
          "workspaceId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "listId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "itemId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "readAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "PageOfNotificationResponse": {
        "required": [
          "value",
          "@odata.nextLink"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            }
          },
          "@odata.nextLink": {
            "type": [
              "null",
              "string"
            ]
          },
          "@odata.count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          }
        },
        "description": "A page of results in Graph/OData shape."
      },
...
      "SettingsRequest": {
        "required": [
          "channels",
          "webhookUrl",
          "quietHoursStart",
          "quietHoursEnd",
          "digestHour"
        ],
        "type": "object",
        "properties": {
          "channels": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "$ref": "#/components/schemas/ChannelChoice"
            }
          },
          "webhookUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "quietHoursStart": {
            "type": [
              "null",
              "string"
            ],
            "format": "time"
          },
          "quietHoursEnd": {
            "type": [
              "null",
              "string"
            ],
            "format": "time"
          },
          "digestHour": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        },
        "description": "Replaces the settings. Channels not listed use the defaults (in-app and webhook on)."
      },
      "SettingsResponse": {
        "required": [
          "channels",
          "webhookUrl",
          "quietHoursStart",
          "quietHoursEnd",
          "digestHour"
        ],
        "type": "object",
        "properties": {
          "channels": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ChannelChoice"
            }
          },
          "webhookUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "quietHoursStart": {
            "type": [
              "null",
              "string"
            ],
            "format": "time"
          },
          "quietHoursEnd": {
            "type": [
              "null",
              "string"
            ],
            "format": "time"
          },
          "digestHour": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "webhookSecret": {
            "type": [
              "null",
              "string"
            ]
          },
          "@odata.etag": {
            "type": [
              "null",
              "string"
            ],
            "description": "The ETag for `If-Match` on changes (the same as the `ETag` header)."
          }
        },
        "description": "Settings; string? SettingsResponse.WebhookSecret is only returned when a secret was created (shown once)."
      },
...
      "UpdateMeRequest": {
        "required": [
          "displayName"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "What users change on their own profile; the e-mail and user name are managed by administrators."
      },
...
    }
  },
...
}